Allow passing a virtual environment to `ruff analyze graph`
Summary
--

Fixes #16598 by adding the `--python` flag to `ruff analyze graph`, which adds a
`PythonPath` to the `SearchPathSettings` for module resolution. For the
[albatross-virtual-workspace] example from the uv repo this updates the output
from the initial issue:

```shell
> ruff analyze graph packages/albatross
{
  "packages/albatross/check_installed_albatross.py": [
    "packages/albatross/src/albatross/__init__.py"
  ],
  "packages/albatross/src/albatross/__init__.py": []
}
```

To include both the the workspace `bird_feeder` import _and_ the third-party
`tqdm` import in the output:

```shell
> myruff analyze graph packages/albatross --python .venv
{
  "packages/albatross/check_installed_albatross.py": [
    "packages/albatross/src/albatross/__init__.py"
  ],
  "packages/albatross/src/albatross/__init__.py": [
    ".venv/lib/python3.12/site-packages/tqdm/__init__.py",
    "packages/bird-feeder/src/bird_feeder/__init__.py"
  ]
}
```

Note the hash in the uv link! I was temporarily very confused why my local tests
were showing an `iniconfig` import instead of `tqdm` until I realized that the
example has been updated on the uv main branch, which I had locally.

Test Plan
--

A new integration test with a stripped down venv based on the `albatross`
example.

[albatross-virtual-workspace]: https://github.com/astral-sh/uv/tree/aa629c4a54c31d6132ab1655b90dd7542c17d120/scripts/workspaces/albatross-virtual-workspace