red_knot_python_semantic: decrease size of `DefinitionKind`
This drops the size from 56 bytes to 24. I made this change on a theory
that this would decrease overall memory usage, since a `DefinitionKind`
seems to be used quite a bit. The theory was that if most values of
this type aren't one of the bigger variants, then most values will end
up using a lot of wasted space.

This didn't seem to end up decreasing overall memory usage in the
tests that I had concocted. However this was kind of an annoying
change, and I don't think there are any downsides as far as I can tell.
Moreover, I think this change is mostly just applying the existing
encapsulation pattern a bit more consistently. And makes potential
future representation changes hopefully a little easier.

If this ended up being proven fruitful, then I would have pursued this
a bit more and decreased its size even more. It would be too hard to
get it down to 16 bytes. But given that I don't have any evidence that
this was actually beneficial, I stopped at this point.