ty_python_semantic: improve diagnostic messages for union type function calls
This PR goes to some effort to improve diagnostics for invalid function
calls, when the type of the function is a union of function types.
Previously, ty was picking the first error and emitting that only. With
this change, we surface all errors. But we try to go about it with some
smarts.

Specifically, when a function call on a union type fails to type
check, we emit the reason that each variant is invalid (which might
only be one) as sub-diagnostics. Moreover, we change the prose on the
diagnostics to connect the initial error message with the added context
in each sub-diagnostic.

One downside here is that this introduces an alternative code path for
each "invalid function call" diagnostic. I resisted the urge to try and
introduce a heavy abstraction here. Instead, we just pass down some
additional data when we're dealing with a union type, and then do case
analysis on that to determine whether, e.g., `INVALID_ARGUMENT_TYPE`
should be its own top-level diagnostic or if it should be a
sub-diagnostic.

I also chose to structure the code how I did to preserve locality. That
is, instead of having two completely different functions for reporting
top-level diagnostics versus sub-diagnostics, I kept everything close
together. So that if you're changing how the top-level
`INVALID_ARGUMENT_TYPE` diagnostic reads, then it should hopefully be
clearer whether the sub-diagnostic needs changing as well.