ty_python_semantic: invalid-union-call sub-diagnostics should respect rule selection
This commit modifies how we add sub-diagnostics for the
invalid-union-call case. Specifically, they now respect rule selection.
So if, for example, `--ignore invalid-argument-type` is given, then that
diagnostic won't show up as a sub-diagnostic of `invalid-union-call`.
And if all of the sub-diagnostics of an `invalid-union-call` are
suppressed, then the overall `invalid-union-call` diagnostic is
suppressed too. Moreover, one can suppress the entire thing regardless
of sub-diagnostics via `--ignore invalid-union-call`.