[ty] Add a warning about pre-release status to the CLI
Summary
--

This was suggested on Discord, I hope this is roughly what we had in mind. I
took the message from the ty README, but I'm more than happy to update it.
Otherwise I just tried to mimic the appearance of the `ruff analyze graph`
warning.

Test Plan
--

New warnings in the CLI tests. I thought this might be undesirable but it looks
like uv did the same thing (https://github.com/astral-sh/uv/pull/6166).