Serialize `Rule` for `CacheMessage` and restore in Pascal-case
Summary
--

This PR is stacked on https://github.com/astral-sh/ruff/pull/18074 and uses the
separation between `CacheMessage` and the other diagnostic types added there to
store a `Rule` directly on `CacheMessage` and convert it back to a `&'static
str` for use by `Diagnostic` and `DiagnosticMessage`.

I think what we really want is to use only the kebab-case names everywhere, but
I don't think we can accomplish that with `Diagnostic::new` living in the
`ruff_diagnostics` crate, and the proc macros being called in `ruff_linter`.
`Diagnostic::new` relies on `ViolationMetadata::rule_name`, which just calls
`stringify!(#name)`.

Test Plan
--

Existing tests