Use static diagnostic names and cache `Rule`s (#18079)
This PR is stacked on https://github.com/astral-sh/ruff/pull/18074 and
uses the
separation between `CacheMessage` and the other diagnostic types added
there to
store a `Rule` directly on `CacheMessage` and convert it back to a
`&'static
str` for use by `Diagnostic` and `DiagnosticMessage`.

In the first commit, I added a generated `Rule::pascal_name` method for
converting
back to the Pascal-case name currently used by `Diagnostic` and
`DiagnosticMessage`,
but in the second commit I used the `heck` crate to generalize our
existing `kebab_case`
macro to convert the `ViolationMetadata::rule_name` output to kebab-case
and use
kebab-case everywhere. `heck` was already in our dependency graph
because it's what
`strum` uses internally for case conversions.

I think that's probably preferable long-term, but it might be a breaking
change in
the LSP because the lint name is used as the title for a quick fix, if a
`suggestion` is
unavailable. Otherwise `Diagnostic::name` is only used for `debug!`
logging.

Test Plan
--

Existing tests